feat: separate service logic from Film controller/storage
