+++ src/main/java/ru/yandex/practicum/filmorate/service/DefaultFilmService.java
@@ -1,0 +1,33 @@
+package ru.yandex.practicum.filmorate.service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
+import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.storage.FilmStorage;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class DefaultFilmService implements FilmService{
+    private final FilmStorage filmStorage;
+
+    public Film addFilm(Film film) {
+        return filmStorage.addFilm(film);
+    }
+
+    public Film updateFilm(Film film) {
+        if (filmStorage.getFilmById(film.getId()) == null) {
+            log.warn("Attempt to update film with unknown id={}:\n{}", film.getId(), film);
+            throw new ValidationException("Фильма с таким ID не зарегистрировано");
+        }
+        return filmStorage.updateFilm(film);
+    }
+
+    public List<Film> getAllFilms() {
+        return filmStorage.getAllFilms();
+    }
+}
