+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,15 +1,13 @@
-package ru.yandex.practicum.filmorate.storage;
+package ru.yandex.practicum.filmorate.service;
 
 import ru.yandex.practicum.filmorate.model.Film;
 
 import java.util.List;
 
-public interface FilmStorage {
+public interface FilmService {
     Film addFilm(Film film);
 
     Film updateFilm(Film film);
 
     List<Film> getAllFilms();
-
-    Film getFilmById(Integer id);
 }
