+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryFilmStorage.java
@@ -2,7 +2,6 @@ package ru.yandex.practicum.filmorate.storage;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
-import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 
 import java.util.ArrayList;
@@ -26,10 +25,6 @@ public class InMemoryFilmStorage implements FilmStorage {
 
     @Override
     public Film updateFilm(Film film) {
-        if (!films.containsKey(film.getId())) {
-            log.warn("Attempt to update film with unknown id={}:\n{}", film.getId(), film);
-            throw new ValidationException("Фильма с таким ID не зарегистрировано");
-        }
         films.put(film.getId(), film);
         log.info("Updated film with id={}:\n{}", film.getId(), film);
         return film;
@@ -38,5 +33,10 @@ public class InMemoryFilmStorage implements FilmStorage {
     @Override
     public List<Film> getAllFilms() {
         return new ArrayList<>(films.values());
+    }
+
+    @Override
+    public Film getFilmById(Integer id) {
+        return films.get(id);
     }
 }
